+++ force_wfmanager/tests/test_wfmanager.py
@@ -1,119 +1,115 @@
-import unittest
-import unittest.mock as mock
-
-from force_wfmanager.wfmanager_plugin import WfManagerPlugin
-
-from envisage.core_plugin import CorePlugin
-from envisage.ui.tasks.tasks_plugin import TasksPlugin
-from force_wfmanager.wfmanager_results_task import WfManagerResultsTask
-from force_wfmanager.wfmanager_setup_task import WfManagerSetupTask
+import contextlib
+import os
+import shutil
+import tempfile
+from unittest import mock, TestCase
 
 from pyface.api import (ConfirmationDialog, YES, NO, CANCEL)
 from pyface.ui.qt4.util.gui_test_assistant import GuiTestAssistant
+from traits.trait_errors import TraitError
 
-from force_bdss.api import (Workflow, WorkflowReader)
-from force_bdss.tests.probe_classes.factory_registry import (
-    ProbeFactoryRegistry
-)
+from force_bdss.api import Workflow
 
-from force_wfmanager.central_pane.analysis_model import AnalysisModel
+from force_wfmanager.model.analysis_model import AnalysisModel
 from force_wfmanager.tests import fixtures
-from force_wfmanager.wfmanager import WfManager, TaskWindowClosePrompt
+from force_wfmanager.wfmanager import TaskWindowClosePrompt
 
-WORKFLOW_READER_PATH = 'force_wfmanager.wfmanager_setup_task.WorkflowReader'
+from .mock_methods import mock_file_reader, mock_dialog
+from .probe_classes import ProbeWfManager
+
+SETUP_ERROR_PATH = 'force_wfmanager.wfmanager_setup_task.error'
+WORKFLOW_READER_PATH = 'force_wfmanager.io.workflow_io.WorkflowReader'
 CONFIRMATION_DIALOG_PATH = 'force_wfmanager.wfmanager.ConfirmationDialog'
 
 
-def dummy_wfmanager(filename=None):
-    plugins = [CorePlugin(), TasksPlugin(),
-               mock_wfmanager_plugin(filename)]
-    wfmanager = WfManager(plugins=plugins)
-    # 'Run' the application by creating windows without an event loop
-    wfmanager.run = wfmanager._create_windows
-    return wfmanager
-
-
-def mock_wfmanager_plugin(filename):
-    plugin = WfManagerPlugin()
-    plugin._create_setup_task = mock_create_setup_task(filename)
-    plugin._create_results_task = mock_create_results_task()
-    return plugin
-
-
-def mock_create_setup_task(filename):
-    def func():
-        wf_manager_task = WfManagerSetupTask(
-            factory_registry=ProbeFactoryRegistry())
-        if filename is not None:
-            wf_manager_task.open_workflow_file(filename)
-        return wf_manager_task
-    return func
-
-
-def mock_create_results_task():
-    def func():
-        wf_manager_task = WfManagerResultsTask(
-            factory_registry=ProbeFactoryRegistry())
-        return wf_manager_task
-    return func
-
-
-def mock_dialog(dialog_class, result, path=''):
-    def mock_dialog_call(*args, **kwargs):
-        dialog = mock.Mock(spec=dialog_class)
-        dialog.open = lambda: result
-        dialog.path = path
-        return dialog
-    return mock_dialog_call
-
-
-def mock_file_reader(*args, **kwargs):
-    def read(*args, **kwargs):
-        workflow = Workflow()
-        return workflow
-    reader = mock.Mock(spec=WorkflowReader)
-    reader.read = read
-    return reader
-
-
-def mock_window(wfmanager):
-    window = mock.Mock(TaskWindowClosePrompt)
-    window.application = wfmanager
-    return window
-
-
-class TestWfManager(GuiTestAssistant, unittest.TestCase):
+class TestWfManager(GuiTestAssistant, TestCase):
     def setUp(self):
         super(TestWfManager, self).setUp()
-        self.wfmanager = dummy_wfmanager()
+        self.wfmanager = ProbeWfManager()
 
+    @contextlib.contextmanager
     def create_tasks(self):
-
         self.wfmanager.run()
         self.setup_task = self.wfmanager.windows[0].tasks[0]
-        self.results_task = self.wfmanager.windows[0].tasks[1]
+        self.review_task = self.wfmanager.windows[0].tasks[1]
+        try:
+            yield
+        finally:
+            for plugin in self.wfmanager:
+                self.wfmanager.remove_plugin(plugin)
+            self.wfmanager.exit()
 
     def test_init(self):
-        self.create_tasks()
-        self.assertEqual(len(self.wfmanager.default_layout), 1)
-        self.assertEqual(len(self.wfmanager.task_factories), 2)
-        self.assertEqual(len(self.wfmanager.windows), 1)
-        self.assertEqual(len(self.wfmanager.windows[0].tasks), 2)
+        with self.create_tasks():
+            self.assertEqual(len(self.wfmanager.default_layout), 1)
+            self.assertEqual(len(self.wfmanager.task_factories), 2)
+            self.assertEqual(len(self.wfmanager.windows), 1)
+            self.assertEqual(len(self.wfmanager.windows[0].tasks), 2)
 
-        self.assertIsInstance(self.setup_task.analysis_model, AnalysisModel)
-        self.assertIsInstance(self.setup_task.workflow_model, Workflow)
+            self.assertIsInstance(
+                self.setup_task.analysis_model, AnalysisModel)
+            self.assertIsInstance(
+                self.setup_task.workflow_model, Workflow)
 
-        self.assertIsInstance(self.results_task.analysis_model, AnalysisModel)
-        self.assertIsInstance(self.results_task.workflow_model, Workflow)
+            self.assertIsInstance(
+                self.review_task.analysis_model, AnalysisModel)
+            self.assertEqual(self.review_task.workflow_model, None)
 
     def test_init_with_file(self):
         with mock.patch(WORKFLOW_READER_PATH) as mock_reader:
             mock_reader.side_effect = mock_file_reader
-            self.wfmanager = dummy_wfmanager(fixtures.get('evaluation-4.json'))
-            self.create_tasks()
-            self.assertEqual(self.setup_task.current_file.split('/')[-1],
-                             'evaluation-4.json')
-            self.assertEqual(mock_reader.call_count, 1)
+            self.wfmanager = ProbeWfManager(
+                fixtures.get('evaluation-4.json'))
+            with self.create_tasks():
+                self.assertEqual(
+                    os.path.basename(self.setup_task.current_file),
+                    'evaluation-4.json')
+                self.assertEqual(mock_reader.call_count, 1)
+
+    def test_init_with_file_failure(self):
+        with mock.patch(SETUP_ERROR_PATH) as mock_error:
+            self.wfmanager = ProbeWfManager("this_file_does_not_exist.json")
+            with self.create_tasks():
+                error_msg = ("Cannot read the requested file:\n\n[Errno 2] "
+                             "No such file or directory: "
+                             "'this_file_does_not_exist.json'")
+                mock_error.assert_called_with(
+                    None,
+                    error_msg,
+                    'Error when reading file'
+                )
+
+    def test_init_with_corrupted_state_file(self):
+        # Add a corrupted applciation_memento to a test location and set
+        # it as a state location
+        temp_dir = tempfile.mkdtemp()
+        self.addCleanup(shutil.rmtree, temp_dir)
+        state_dir = os.path.join(temp_dir, "example_state_location")
+        os.mkdir(state_dir)
+        ref_state_file = os.path.join(
+            fixtures.get('example_state_location'),
+            "application_memento.CORRUPTED")
+        target_state_file = os.path.join(state_dir, "application_memento")
+        shutil.copyfile(ref_state_file, target_state_file)
+        self.wfmanager.state_location = state_dir
+
+        with mock.patch('force_wfmanager.wfmanager.log') as mock_log:
+            try:
+                self.wfmanager.run()
+                self.setup_task = self.wfmanager.windows[0].tasks[0]
+                self.review_task = self.wfmanager.windows[0].tasks[1]
+            except TraitError:
+                self.fail("Error: did the corrupted state file make "
+                          "its way through?")
+            finally:
+                # cleanup
+                for plugin in self.wfmanager:
+                    self.wfmanager.remove_plugin(plugin)
+                self.wfmanager.exit()
+
+            mock_log.warning.assert_called_once_with(
+                'The state file at {!r} was corrupted and has been removed.'
+                .format(target_state_file))
 
     def test_remove_tasks_on_application_exiting(self):
         self.wfmanager.run()
@@ -122,46 +118,52 @@ class TestWfManager(GuiTestAssistant, unittest.TestCase):
         self.assertEqual(len(self.wfmanager.windows[0].tasks), 0)
 
     def test_switch_task(self):
-        self.create_tasks()
-        self.assertEqual(
-            self.wfmanager.windows[0].active_task,
-            self.setup_task,
-            msg='Note: this test can fail locally if a saved application '
-                'memento exists with the results task in focus'
-        )
-        self.wfmanager.windows[0].active_task.switch_task()
-        self.assertEqual(self.wfmanager.windows[0].active_task,
-                         self.results_task)
-        self.wfmanager.windows[0].active_task.switch_task()
-        self.assertEqual(self.wfmanager.windows[0].active_task,
-                         self.setup_task)
+        with self.create_tasks():
+            self.assertEqual(
+                self.wfmanager.windows[0].active_task,
+                self.setup_task,
+                msg='Note: this test can fail locally if a saved application '
+                    'memento exists with the review task in focus'
+            )
+            self.wfmanager.windows[0].active_task.switch_task()
+            self.assertEqual(self.wfmanager.windows[0].active_task,
+                             self.review_task)
+            self.wfmanager.windows[0].active_task.switch_task()
+            self.assertEqual(self.wfmanager.windows[0].active_task,
+                             self.setup_task)
 
     def test_result_task_exit(self):
-        self.create_tasks()
-        for window in self.wfmanager.windows:
-            window.close = mock.Mock(return_value=True)
-        self.results_task.exit()
-        self.assertTrue(self.results_task.window.close.called)
+        with self.create_tasks():
+            for window in self.wfmanager.windows:
+                window.close = mock.Mock(return_value=True)
+            self.review_task.exit()
+            self.assertTrue(self.review_task.window.close.called)
 
     def test_setup_task_exit(self):
-        self.create_tasks()
-        for window in self.wfmanager.windows:
-            window.close = mock.Mock(return_value=True)
-        self.setup_task.exit()
-        self.assertTrue(self.setup_task.window.close.called)
+        with self.create_tasks():
+            for window in self.wfmanager.windows:
+                window.close = mock.Mock(return_value=True)
+            self.setup_task.exit()
+            self.assertTrue(self.setup_task.window.close.called)
 
 
-class TestTaskWindowClosePrompt(unittest.TestCase):
+class TestTaskWindowClosePrompt(TestCase):
 
     def setUp(self):
         super(TestTaskWindowClosePrompt, self).setUp()
-        self.wfmanager = dummy_wfmanager()
+        self.wfmanager = ProbeWfManager()
         self.create_tasks()
+
+    def tearDown(self):
+        # In case a faulty test hasn't succeeded closing via the dialog
+        for plugin in self.wfmanager:
+            self.wfmanager.remove_plugin(plugin)
+        self.wfmanager.exit()
 
     def create_tasks(self):
         self.wfmanager.run()
         self.setup_task = self.wfmanager.windows[0].tasks[0]
-        self.results_task = self.wfmanager.windows[0].tasks[1]
+        self.review_task = self.wfmanager.windows[0].tasks[1]
 
     def test_exit_application_with_saving(self):
         self.setup_task.save_workflow = mock.Mock(return_value=True)